Write tests for encryption and decryption with hmac-secret included.
XXX These tests rely on reaching into the internal ABI of libfido2 to
mock setting the hmac-secret.  This is a horrible kludge, but I don't
see any other way to mock it up -- not even clear it's possible to
mock up the ECDH key agreement in the authenticator/host interaction.
This kludge should be expunged as soon as there is a better way:

https://github.com/Yubico/libfido2/issues/256