Fix bugs with short coordinates.
Both es256_pk_from_EC_KEY from libfido2 until 1.11.0, and our local
version of es256_pk_encode, used BN_bn2bin to encode an OpenSSL
BIGNUM in big-endian -- which always puts the highest-order _nonzero_
byte in the first byte position of the output.

When big-endian encoding is narrower than a fixed-width output
buffer, this has the effect of multiplying it by 256, which is very
much not useful!

For example, encoding 0x1a2b3c4d in a 4-byte big-endian buffer gives:

        1a 2b 3c 4d

Encoding 0x1a2b3c = 0x001a2b3c in a 4-byte big-endian buffer _should_
give:

        00 1a 2b 3c

But (with a zero-initialized output buffer), the logic that used
BN_bn2bin without adjusting the first byte position according to the
number of bytes in the output would encode it as:

        1a 2b 3c 00

which, when read back, becomes 0x1a2b3c00 = 256*0x1a2b3c.

The practical impact of this is that for some public keys (around
1/128 of them), a round-trip through encoding and decoding would
yield garbled results or fail altogether because the coordinates give
a point that is not on the curve.

=> Applications which, on registration, use both fido_cred_encrypt
   and then fido_assert_decrypt would immediately fail.  For example,
   fidocrypt(1) does (and already did) this, so there is no danger it
   might have recorded unrecoverable credentials.  (This is now
   recommended in the man page.)

=> Applications which, on registration, use only fido_cred_encrypt
   might have stored ciphertexts that can't be decrypted.

Either adjusting the starting byte according to the position of the
highest-order byte, or, better, using the more-likely constant-time
BN_bn2binpad, resolves the issue.  This is what we do now in
es256_pk_encode.

The bug in libfido2 es256_pk_from_EC_KEY has been fixed upstream in
1.11.0.  To work around this with older libfido2, we define a local
fidocrypt_es256_pk_from_EC_KEY that uses BN_bn2binpad.

New test cases for fido_cred_encrypt and fido_assert_decrypt detect
this situation in both directions by using a public key where at
least one of the coordinates of a point on the NIST P-256 curve has
fewer than 31 bytes.