fidocrypt(1): Limit threads to 64 KB stacks.
Kind of arbitrary depth is probably enough for a call into libfido2
to get a credential or assertion or a call into sqlite3 to query a
fidocrypt file.

Measurements with mincore show <=8 KB on NetBSD, so 64 KB has ample
margin for much more.

mlock forces the stacks to be preallocated in RAM, so it's prudent to
keep them small.